finanzas-mcp-server/pyproject.toml
[project]
name = "finanzas-mcp-server"
version = "0.1.0"
description = "MCP Server for Finanzas Personales - NL to SQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # MCP Server
    "mcp[cli]>=1.0.0",
    
    # Database
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    
    # Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    
    # OpenAI (embeddings + LLM)
    "openai>=1.50.0",
    
    # Qdrant (vector storage)
    "qdrant-client>=1.12.0",
    
    # SQL Parsing
    "sqlparse>=0.5.0",
    
    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
finanzas-mcp = "src.server:main"
